prod-requirements.txt
alembic==1.2.1
aniso8601==8.0.0
arrow==0.15.2
Babel==2.9.0
beautifulsoup4==4.6.0
blinker==1.4
certifi==2019.9.11
cffi==1.14.5
chardet==3.0.4
Click==7.0
cryptography==3.4.7
decorator==4.4.0
dnspython==2.1.0
dominate==2.4.0
email-validator==1.1.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-Security-Too==3.4.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
grako==3.99.9
idna==2.8
infinity==1.4
intervals==0.8.1
invoke==1.3.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
line-bot-sdk==1.19.0
lml==0.0.9
Mako==1.1.0
MarkupSafe==1.1.1
msal==1.10.0
numpy==1.17.2
openpyxl==2.5.3
pandas==0.25.1
pandocfilters==1.4.2
passlib==1.7.4
psycopg2==2.9.3
pycparser==2.20
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.7
PyJWT==2.0.1
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-download==0.1.2
rq==1.1.0
rq-dashboard==0.5.3
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.7
urllib3==1.25.6
validators==0.14.0
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
WTForms-Components==0.10.4
WTForms-JSON==0.3.3
XlsxWriter==1.2.1
